Added slow rotation (shift+left/right)

+++ helperfunctions.cpp
@@ -52,12 +52,12 @@ double sine(double degrees)
     return std::sin(degrees * (bmc::two_pi<double>() / 360.0));
 }
 
-int sgn(int x)
+int sgn(double x)
 {
-    if (x > 0) {
+    if (x > 0.0) {
         return 1;
     }
-    if (x < 0) {
+    if (x < 0.0) {
         return -1;
     }
     return 0;
